Quick restructuring SearchScreen.kt
