+++ androidApp/src/main/java/com/rld/justlisten/android/ui/searchscreen/components/ShowPreviousSearches.kt
@@ -1,0 +1,11 @@
+package com.rld.justlisten.android.ui.searchscreen.components
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.util.fastForEach
+
+@Composable
+fun ShowPreviousSearches(listOfSearches: List<String>, onPreviousSearchedPressed: (String) ->Unit) {
+    listOfSearches.fastForEach { itemSearched ->
+        ItemRowSearch(itemSearched, onPreviousSearchedPressed)
+    }
+}
